fix: adjust footer and recheck nested error button
